+++ ChangeLog.md
@@ -6,7 +6,8 @@ WebSockets change log
 * **Heads up**: Deprecated passing origin to `WebSocket` constructor. It
   should be passed inside the headers when calling *connect()*.
   (@thekid)
-* Added ability to pass path and query string to `WebSocket` constructor
+* Merged PR #7: Added ability to pass path and query string to `WebSocket`
+  constructor
   (@thekid)
 * Fixed "Call to a member function message() on null" errors when using
   an already connected socket in the `WebSocket` constructor.